fixec cloudinary.js,user.routes.js,user.controller.js bugs,   wrote multer.middleware.js

+++ src/utils/cloudinary.js
@@ -1,28 +1,42 @@
 import {v2 as cloudinary} from 'cloudinary';
 import fs from "fs";
+// import dotenv from "dotenv";
+
+// dotenv.config({
+//     path: './.env'
+// })
 
 cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+        cloud_name:process.env.CLOUDINARY_CLOUD_NAME, 
         api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
     });
 
     const uploadOnCloudinary=async (localFilePath) => {
         try{
             if(!localFilePath) return null;
             const response =await cloudinary.uploader.upload(localFilePath,{
-                resource_type:auto
+                resource_type:"auto"
             })
+            console.log("response dedo bhai",response);
 
             //file has been uploaded successfull
             console.log("file is uploaded on cloudinary",
             response.url);
             return response;
         }catch(error){
-            fs.unlinkSync(localFilePath) // remove the locally saved temporary  file as the upload 
+                console.error("Cloudinary upload error:", error);
+                
+                // Safely unlink if file exists
+                if (fs.existsSync(localFilePath)) {
+                fs.unlinkSync(localFilePath);
+                }
+                
+                return null; // ✅ Return null if upload failed // remove the locally saved temporary  file as the upload 
         }
     }
 
+    export {uploadOnCloudinary};
 // Upload an image
 //         const uploadResult = await cloudinary.uploader
 //         .upload(
